% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaNorm.R
\name{box_cox_transformation}
\alias{box_cox_transformation}
\title{Box-Cox Transformation}
\usage{
box_cox_transformation(data)
}
\arguments{
\item{data}{A single variable.}
}
\description{
Compute Box-Cox normalization.
}
