% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaCorr.R
\name{cor_pmat}
\alias{cor_pmat}
\title{Correlation P-Values}
\usage{
cor_pmat(x, method, ...)
}
\arguments{
\item{x}{A data matrix.}

\item{method}{Character indicating which correlation coefficient has to be computed. Options are "pearson" (default), "kendall" and "spearman".}
}
\description{
Compute correlation p-values.
}
