% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaSummarizedExperiment.R
\name{detect_decimals}
\alias{detect_decimals}
\title{Detect decimals}
\usage{
detect_decimals(data)
}
\arguments{
\item{data}{A data matrix (samples in rows).}
}
\description{
Detect decimal variables.
}
