% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaCorr.R
\name{flattenCorrMatrix}
\alias{flattenCorrMatrix}
\title{Flatten Correlation Matrix}
\usage{
flattenCorrMatrix(cormat, pmat)
}
\arguments{
\item{cormat}{Output from \code{cor}.}

\item{pmat}{Output from \code{cor_pmat}.}
}
\description{
Flatten Correlation Matrix
}
