% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{poma_pal_c}
\alias{poma_pal_c}
\title{Return function to interpolate a continuous POMA color palette}
\usage{
poma_pal_c(palette = "nature")
}
\arguments{
\item{palette}{Character name of palette in poma_palettes}
}
\description{
Return function to interpolate a continuous POMA color palette
}
