% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{poma_pal_d}
\alias{poma_pal_d}
\title{Return function to interpolate a discrete POMA color palette}
\usage{
poma_pal_d(palette = "nature")
}
\arguments{
\item{palette}{Character name of palette in poma_palettes}
}
\description{
Return function to interpolate a discrete POMA color palette
}
