% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaCorr.R
\name{PomaCorr}
\alias{PomaCorr}
\title{Correlation Analysis}
\usage{
PomaCorr(data, method = "pearson", label_size = 8, theme_params = list())
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{method}{Character. Indicates which correlation coefficient has to be computed. Options are "pearson" (default), "kendall" and "spearman".}

\item{label_size}{Numeric. Indicates plot label size.}

\item{theme_params}{List. Indicating \code{theme_poma} parameters.}
}
\value{
A \code{list} with the results.
}
\description{
\code{PomaCorr} computes all pairwise correlations in the data.
}
\examples{
## Output is a list with objects `correlations` (tibble) and `corrplot` (ggplot2 object)
data <- POMA::st000284 # Example SummarizedExperiment object included in POMA

data \%>\% 
  PomaCorr(method = "pearson")
}
\references{
Jerome Friedman, Trevor Hastie and Rob Tibshirani (2019). glasso: Graphical Lasso: Estimation of Gaussian Graphical Models. R package version 1.11. https://CRAN.R-project.org/package=glasso
}
\author{
Pol Castellano-Escuder
}
