% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaDESeq.R
\name{PomaDESeq}
\alias{PomaDESeq}
\title{Differential Expression Analysis Based on the Negative Binomial Distribution}
\usage{
PomaDESeq(data, adjust = "fdr")
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{adjust}{Character. Multiple comparisons correction method to adjust p-values. Available options are: "fdr" (false discovery rate), "holm", "hochberg", "hommel", "bonferroni", "BH" (Benjamini-Hochberg), and "BY" (Benjamini-Yekutieli).}
}
\value{
A \code{tibble} with the results.
}
\description{
\code{PomaDESeq} estimates variance-mean dependence in count data from high-throughput sequencing assays and test for differential expression based on a model using the negative binomial distribution.
}
\references{
Love, M.I., Huber, W., Anders, S. Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2 Genome Biology 15(12):550 (2014)
}
\author{
Pol Castellano-Escuder
}
