% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaEDA.R
\name{PomaEDA}
\alias{PomaEDA}
\title{Automatic Exploratory Data Analysis PDF Report}
\usage{
PomaEDA(
  data,
  imputation = "knn",
  normalization = "log_pareto",
  clean_outliers = TRUE,
  coeff_outliers = 1.5,
  username = "Username"
)
}
\arguments{
\item{data}{A SummarizedExperiment object. First \code{colData} column must be the subject group/type.}

\item{imputation}{Imputation method. Options are "none", "half_min", "median", "mean", "min" and "knn" (default). If "none", all missing values will be replaced by zero.}

\item{normalization}{Normalization method. Options are "none", "auto_scaling", "level_scaling", "log_scaling", "log_transformation", "vast_scaling" and "log_pareto" (default).}

\item{clean_outliers}{Logical. If it's set to TRUE, outliers will be removed from EDA.}

\item{coeff_outliers}{This value corresponds to the classical 1.5 in \eqn{Q3 + 1.5*IQR} formula to detect outliers. By changing this value, the permissiveness in outlier detection will change.}

\item{username}{This name will be included as a report subtitle.}
}
\value{
An exploratory data analysis PDF report.
}
\description{
This function automatically generates a PDF report with different exploratory plots and tables from an SummarizedExperiment object.
}
\author{
Pol Castellano-Escuder
}
