% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaHeatmap.R
\name{PomaHeatmap}
\alias{PomaHeatmap}
\title{Classical Heatmap}
\usage{
PomaHeatmap(
  data,
  cols = 1,
  sample_names = TRUE,
  feature_names = FALSE,
  show_legend = TRUE
)
}
\arguments{
\item{data}{A SummarizedExperiment object. First \code{colData} column must be the subject group/type.}

\item{cols}{Numerical vector indicating the column index of variables in \code{colData} to be displayed. Default is 1 (main group).}

\item{sample_names}{Logical indicating if sample names should be plotted or not. Default is TRUE.}

\item{feature_names}{Logical indicating if feature names should be plotted or not. Default is FALSE.}

\item{show_legend}{Logical indicating if legend should be plotted or not. Default is TRUE.}
}
\value{
A heatmap plot.
}
\description{
This function returns a basic heatmap plot made with base R.
}
\examples{
data("st000284")

st000284 \%>\% 
  PomaNorm() \%>\% 
  PomaHeatmap()
}
\author{
Pol Castellano-Escuder
}
