% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaMSnSetClass.R
\name{PomaMSnSetClass}
\alias{PomaMSnSetClass}
\title{Convert data frames to a MSnSet Object}
\usage{
PomaMSnSetClass(target, features)
}
\arguments{
\item{target}{Metadata variables structured in columns. Sample ID must be the first column and group/type/treatment of the study must be the second column.}

\item{features}{Table of features. Each feature in one column.}
}
\value{
A MSnSet object.
}
\description{
This function converts data frame objects to a MSnSet object.
}
\examples{
data(iris)

# create target: two column (or more) data frame with IDs and Group factor
target <- data.frame(ID = 1:150, Group = iris$Species)

# create features: p column data frame (or matrix) with features
features <- iris[,1:4]

# create an MSnSet object with POMA
object <- PomaMSnSetClass(target = target, features = features)
}
\references{
Laurent Gatto and Kathryn S. Lilley. MSnbase - an R/Bioconductor package for isobaric tagged mass spectrometry data visualization, processing and quantitation. Bioinformatics 28, 288-289 (2012).
}
\author{
Pol Castellano-Escuder
}
