% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaNormPlot.R
\name{PomaNormPlot}
\alias{PomaNormPlot}
\title{Normalization Plot}
\usage{
PomaNormPlot(data, group = c("subjects", "metabolites"))
}
\arguments{
\item{data}{A data frame with metabolites. First column must be the subject ID and second column must be a factor with the subject group.}

\item{group}{Groupping factor for the plot. Options are c("subjects", "metabolites"). If the user select "subject", the boxplot will be created for each subject. If the selection is "metabolites", the boxplot will be created for each metabolite.}
}
\value{
A ggplot2 object.
}
\description{
PomaNormPlot() generates a boxplot of not normalized and normalized metabolomic data. This plot can help in the comparison between pre and post normalized data and in the "validation" of the normalization process.
}
\author{
Pol Castellano-Escuder
}
