% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaOddsRatio.R
\name{PomaOddsRatio}
\alias{PomaOddsRatio}
\title{Logistic Regression Model Odds Ratios}
\usage{
PomaOddsRatio(data, feature_name = NULL)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{feature_name}{A vector with the name/s of feature/s to plot. If it's NULL (default), all variables will be plotted.}
}
\value{
A data frame with the Odds Ratios for all features with their 95\% confidence intervals and a ggplot2 object.
}
\description{
PomaOddsRatio() calculates the Odds Ratios for each feature from a logistic regression model using the binary outcone (group/type must be a binary factor) as a dependent variable.
}
\author{
Pol Castellano-Escuder
}
