% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaOutliers.R
\name{PomaOutliers}
\alias{PomaOutliers}
\title{Remove and Analyze Outliers}
\usage{
PomaOutliers(
  data,
  do = "clean",
  method = "euclidean",
  type = "median",
  iqr = 1.5,
  labels = FALSE
)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{do}{Action to do. Options are "clean" (to remove detected outliers) and "analyze" (to analyze data outliers). Note that the output of this function will be different depending on this parameter.}

\item{method}{blabla}

\item{type}{blabla}

\item{iqr}{blabla}

\item{labels}{blabla}
}
\value{
A MSnSet object with cleaned data or different plots for the detailed analysis of outliers (depending on "do" parameter).
}
\description{
blabla
}
\author{
Pol Castellano-Escuder
}
