% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaPLS.R
\name{PomaPLS}
\alias{PomaPLS}
\title{Partial Least Squares Methods}
\usage{
PomaPLS(
  data,
  method = "pls",
  y = NULL,
  ncomp = 5,
  labels = FALSE,
  ellipse = TRUE,
  validation = "Mfold",
  folds = 5,
  nrepeat = 10,
  vip = 1,
  num_features = 10,
  theme_params = list(),
  ...
)
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{method}{Character. PLS method. Options include "pls", "plsda", and "splsda".}

\item{y}{Character. Indicates the name of \code{colData} columns to be used as dependent variable. If it's set to NULL, the first variable in \code{colData} will be used as the dependent variable.}

\item{ncomp}{Numeric. Number of components in the model. Default is 5.}

\item{labels}{Logical. Indicates if sample names should be displayed.}

\item{ellipse}{Logical. Indicates whether a 95 percent confidence interval ellipse should be displayed. Default is TRUE.}

\item{vip}{Numeric. (Only for "plsda" method). Indicates the variable importance in the projection (VIP) cutoff.}

\item{num_features}{Numeric. (Only for "splsda" method). Number of features to discriminate groups.}

\item{theme_params}{List. Indicates \code{theme_poma} parameters.}
}
\value{
A \code{list} with results including plots and tables.
}
\description{
\code{PomaPLS} performs Partial Least Squares (PLS) regression, Partial Least Squares Discriminant Analysis (PLS-DA) to classify samples, and Sparse Partial Least Squares Discriminant Analysis (sPLS-DA) to classify samples (supervised analysis) and select variables.
}
\examples{
data("st000284")

# PLS
st000284 \%>\% 
  PomaNorm() \%>\%
  PomaPLS(method = "pls")
  
data("st000336")

# PLSDA
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaPLS(method = "plsda")

# sPLSDA
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaPLS(method = "splsda")
}
\author{
Pol Castellano-Escuder
}
