% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaRandForest.R
\name{PomaRandForest}
\alias{PomaRandForest}
\title{Classification Random Forest for Mass Spectrometry Data}
\usage{
PomaRandForest(
  data,
  folds = 3,
  ntree = 500,
  mtry = floor(sqrt(ncol(t(Biobase::exprs(data))))),
  nodesize = 1,
  nvar = 20
)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the suject group/type.}

\item{folds}{Number of observations that will be used as test dataset. For example, if folds = 3, 1/3 of dataset will be used as train dataset.}

\item{ntree}{Number of trees to grow.}

\item{mtry}{Number of variables randomly sampled as candidates at each split. This value is set sqrt(p) (where p is number of variables in data) by default.}

\item{nodesize}{Minimum size of terminal nodes. By default is equal to 1.}

\item{nvar}{Number of variables to show in the Gini plot.}
}
\value{
A list with all results for Random Forest including plots and data frames.
}
\description{
PomaRandForest() allows users to perform a classification Random Forest with a MS data matrix using the classical \code{randomForest} R package.
}
\references{
A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. R News 2(3), 18--22.
}
\author{
Pol Castellano-Escuder
}
