% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaRankProd.R
\name{PomaRankProd}
\alias{PomaRankProd}
\title{Rank Product/Rank Sum Analysis for Mass Spectrometry Data}
\usage{
PomaRankProd(
  data,
  logged = TRUE,
  logbase = 2,
  paired = NA,
  cutoff = 0.05,
  method = "pfp"
)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{logged}{If "TRUE" (default) data have been previously log transformed.}

\item{logbase}{Numerical. Base for log transformation.}

\item{paired}{Number of random pairs generated in the function, if set to NA (default), the odd integer closer to the square of the number of replicates is used.}

\item{cutoff}{The pfp/pvalue threshold value used to select features.}

\item{method}{If cutoff is provided, the method needs to be selected to identify features. "pfp" uses percentage of false prediction, which is a default setting. "pval" uses p-values which is less stringent than pfp.}
}
\value{
A list with all results for Rank Product analysis including tables and plots.
}
\description{
PomaRankProd() performs the Rank Product method to identify differential feature concentration/intensity.
}
\references{
Breitling, R., Armengaud, P., Amtmann, A., and Herzyk, P.(2004) Rank Products: A simple, yet powerful, new method to detect differentially regulated genes in replicated microarray experiments, FEBS Letter, 57383-92

Hong, F., Breitling, R., McEntee, W.C., Wittner, B.S., Nemhauser, J.L., Chory, J. (2006). RankProd: a bioconductor package for detecting differentially expressed genes in meta-analysis Bioinformatics. 22(22):2825-2827

Del Carratore, F., Jankevics, A., Eisinga, R., Heskes, T., Hong, F. & Breitling, R. (2017). RankProd 2.0: a refactored Bioconductor package for detecting differentially expressed features in molecular profiling datasets. Bioinformatics. 33(17):2774-2775
}
\author{
Pol Castellano-Escuder
}
