% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaSummarizedExperiment.R
\name{PomaSummarizedExperiment}
\alias{PomaSummarizedExperiment}
\title{Convert data frames to a SummarizedExperiment Object}
\usage{
PomaSummarizedExperiment(target, features)
}
\arguments{
\item{target}{Metadata variables structured in columns. Sample ID must be the first column.}

\item{features}{Matrix of features. Each feature in one column.}
}
\value{
A SummarizedExperiment object.
}
\description{
This function converts data frame objects to a SummarizedExperiment object.
}
\examples{
data(iris)

# create target: two column (or more) data frame with IDs and Group factor
target <- data.frame(ID = 1:150, Group = iris$Species)

# create features: p column data frame (or matrix) with features
features <- iris[,1:4]

# create an SummarizedExperiment object with POMA
object <- PomaSummarizedExperiment(target = target, features = features)
}
\references{
Morgan M, Obenchain V, Hester J, Pagès H (2021). SummarizedExperiment: SummarizedExperiment container. R package version 1.24.0, https://bioconductor.org/packages/SummarizedExperiment.
}
\author{
Pol Castellano-Escuder
}
