% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaUMAP.R
\name{PomaUMAP}
\alias{PomaUMAP}
\title{Dimensionality Reduction with UMAP}
\usage{
PomaUMAP(data)
}
\arguments{
\item{data}{A SummarizedExperiment object.}
}
\value{
A list with results.
}
\description{
Carry out dimensionality reduction of a dataset using the Uniform Manifold Approximation and Projection (UMAP) method.
}
\examples{
data("st000284")

st000284 \%>\%
  PomaNorm() \%>\%
  PomaUMAP()
}
\references{
McInnes, L., Healy, J., & Melville, J. (2018). Umap: Uniform manifold approximation and projection for dimension reduction. arXiv preprint arXiv:1802.03426.
}
\author{
Pol Castellano-Escuder
}
