% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_poma_c}
\alias{scale_color_poma_c}
\title{Color scale constructor for continuous \code{viridis} "plasma" palette}
\usage{
scale_color_poma_c()
}
\description{
Color scale constructor for continuous \code{viridis} "plasma" palette
}
