% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_poma_d}
\alias{scale_color_poma_d}
\title{Color scale constructor for discrete \code{viridis} "plasma" palette}
\usage{
scale_color_poma_d()
}
\description{
Color scale constructor for discrete \code{viridis} "plasma" palette
}
