% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_poma_c}
\alias{scale_fill_poma_c}
\title{Fill scale constructor for continuous POMA color palettes}
\usage{
scale_fill_poma_c(palette = "nature")
}
\arguments{
\item{palette}{Character name of palette in poma_palettes}
}
\description{
Fill scale constructor for continuous POMA color palettes
}
