% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_poma_d}
\alias{scale_fill_poma_d}
\title{Fill scale constructor for discrete \code{viridis} "plasma" palette}
\usage{
scale_fill_poma_d()
}
\description{
Fill scale constructor for discrete \code{viridis} "plasma" palette
}
