% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st000336}
\alias{st000336}
\title{Targeted LC/MS of urine from boys with DMD and controls}
\format{
A \code{MSnSet} object: 57 samples, 31 metabolites, 1 covariable and 2 groups (Controls and DMD).
\describe{
\item{metabolites}{31 urine metabolites.}
\item{covariables}{Steroid status.}
}
}
\source{
\url{https://www.metabolomicsworkbench.org/data/DRCCMetadata.php?Mode=Study&DataMode=AllData&StudyID=ST000336&StudyType=MS&ResultType=1#DataTabs}
}
\usage{
st000336
}
\description{
Duchenne Muscular Dystrophy (DMD) is an X-linked recessive form of muscular dystrophy that affects males via a mutation in the
gene for the muscle protein, dystrophin. Progression of the disease results in severe muscle loss, ultimately leading to paralysis
and death. Steroid therapy has been a commonly employed method for reducing the severity of symptoms. This study aims to quantify
the urine levels of amino acids and organic acids in patients with DMD both with and without steroid treatment. Track the progression
of DMD in patients who have provided multiple urine samples.
}
\keyword{datasets}
