% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaNorm.R
\name{sum_norm}
\alias{sum_norm}
\title{Sample Sum Normalization}
\usage{
sum_norm(data)
}
\arguments{
\item{data}{A data matrix (samples in rows).}
}
\description{
Compute sum normalization. Final unit is a percentage.
}
