% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaBoxplots.R
\name{PomaBoxplots}
\alias{PomaBoxplots}
\title{Classical Boxplots}
\usage{
PomaBoxplots(
  data,
  group = "samples",
  jitter = TRUE,
  feature_name = NULL,
  label_size = 10
)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{group}{Groupping factor for the plot. Options are "samples" and "features". Option "samples" (default) will create a boxplot for each sample and option "features" will create a boxplot of each variable.}

\item{jitter}{Logical. If it's TRUE (default), the boxplot will show all points.}

\item{feature_name}{A vector with the name/s of feature/s to plot. If it's NULL (default) a boxplot of all features will be created.}

\item{label_size}{Numeric indicating the size of x-axis labels.}
}
\value{
A ggplot2 object.
}
\description{
PomaBoxplots() generates a boxplot for subjects or features. This boxplot can help in the comparison between pre and post normalized data and in the "validation" of the normalization process.
}
\author{
Pol Castellano-Escuder
}
