% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaBoxplots.R
\name{PomaBoxplots}
\alias{PomaBoxplots}
\title{Classical Boxplots}
\usage{
PomaBoxplots(
  data,
  group = "samples",
  jitter = FALSE,
  feature_name = NULL,
  theme_params = list(legend_title = FALSE, axis_x_rotate = TRUE),
  palette = "nature",
  ...
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{group}{Grouping factor for the plot. Options are "samples" and "features". Option "samples" (default) will create a boxplot for each sample and option "features" will create a boxplot of each variable.}

\item{jitter}{Logical. If it's TRUE (default), the boxplot will show all points.}

\item{feature_name}{A vector with the name/s of feature/s to plot. If it's NULL (default) a boxplot of all features will be created.}

\item{theme_params}{List indicating \code{theme_poma} parameters.}

\item{palette}{POMA palette. One of "nature", "simpsons", or "futurama".}
}
\value{
A \code{ggplot} object.
}
\description{
PomaBoxplots() generates a boxplot for subjects or features. This boxplot can help in the comparison between pre and post normalized data and in the "validation" of the normalization process.
}
\examples{
data("st000284")

# samples
st000284 \%>\%
PomaNorm() \%>\% 
PomaBoxplots(theme_params = list(axistext = "y"))

# features
st000284 \%>\% 
PomaNorm() \%>\% 
PomaBoxplots(group = "features", theme_params = list(axis_x_rotate = TRUE))
             
# concrete features
PomaBoxplots(st000284, group = "features", 
             feature_name = c("ornithine", "orotate"))
}
\author{
Pol Castellano-Escuder
}
