% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaClust.R
\name{PomaClust}
\alias{PomaClust}
\title{Cluster Analysis}
\usage{
PomaClust(
  data,
  method = "euclidean",
  k = 3,
  show_clusters = TRUE,
  show_labels = FALSE
)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{method}{Distance measure method to perform MDS. Options are "euclidean", "maximum", "manhattan", "canberra", "binary" and "minkowski". See \code{?dist()}.}

\item{k}{Number of clusters.}

\item{show_clusters}{Logical indicating if clusters should be plotted or not. If this parameter is set to \code{FALSE} the resultant plot will be a classical 2-dimension MDS plot.}

\item{show_labels}{Logical indicating if sample names should be plotted or not.}
}
\value{
A list with the results.
}
\description{
This function performs a classical multidimensional scaling (MDS) using all features in the data and computes a cluster analysis for \code{k} clusters. Then, the calculated clusters will be represented on a MDS plot.
}
\author{
Pol Castellano-Escuder
}
