% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaClust.R
\name{PomaClust}
\alias{PomaClust}
\title{Cluster Analysis}
\usage{
PomaClust(
  data,
  method = "euclidean",
  k = NA,
  k_max = 15,
  show_clusters = TRUE,
  labels = FALSE,
  show_group = FALSE
)
}
\arguments{
\item{data}{A SummarizedExperiment object. First \code{colData} column must be the subject group/type.}

\item{method}{Distance measure method to perform MDS. Options are "euclidean", "maximum", "manhattan", "canberra" and "minkowski". See \code{?dist()}.}

\item{k}{Number of clusters (default is \code{NA}). The optimum number of clusters will be used by default.}

\item{k_max}{Number of clusters among which the optimal one will be selected.}

\item{show_clusters}{Logical indicating if clusters should be plotted or not. If this parameter is set to \code{FALSE} the resultant plot will be a classical 2-dimension MDS plot.}

\item{labels}{Logical indicating if sample names should be plotted or not.}

\item{show_group}{Logical indicating if the original sample group from pData should be plotted instead of sample ID or not. Only works if labels is set to \code{TRUE}.}
}
\value{
A list with the results.
}
\description{
This function performs a classical multidimensional scaling (MDS) using all features in the data and computes a cluster analysis for \code{k} clusters. Then, the calculated clusters will be represented on a MDS plot.
}
\examples{
data("st000284")

PomaClust(st000284)
}
\author{
Pol Castellano-Escuder
}
