% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaCorr.R
\name{PomaCorr}
\alias{PomaCorr}
\title{Correlation Analysis}
\usage{
PomaCorr(
  data,
  method = "pearson",
  low = "red",
  outline = "white",
  high = "blue",
  label_size = 12,
  corr_type = "cor",
  coeff = 0.7
)
}
\arguments{
\item{data}{A SummarizedExperiment object. First \code{colData} column must be the subject group/type.}

\item{method}{Character indicating which correlation coefficient has to be computed. Options are "pearson" (default), "kendall" and "spearman".}

\item{low}{Color for low end of the gradient in corrplot.}

\item{outline}{Color for the outline of the gradient in corrplot.}

\item{high}{Color for high end of the gradient in corrplot.}

\item{label_size}{Numeric indicating label size in corrplot.}

\item{corr_type}{Type of correlation network. Options are "cor" (for global correlations) and "glasso" (for gaussian graphical model). Default is "cor". See \code{glasso} R package for the second option.}

\item{coeff}{Numeric indicating correlation coefficient. Edges with absolute weight below this value will be removed from the network. If "corr_type" is set to "glasso", this parameter indicates the regularization parameter for lasso (rho = 0 means no regularization). See \code{glasso::glasso()}.}
}
\value{
A list with the results.
}
\description{
This function returns different correlation plots and a table with all pairwise correlations in the data.
}
\examples{
data("st000284")

# Pearson correlation
PomaCorr(st000284)$correlations

## Gaussian graphical model
# library(ggraph)
# PomaCorr(st000284, corr_type = "glasso")
}
\references{
Jerome Friedman, Trevor Hastie and Rob Tibshirani (2019). glasso: Graphical Lasso: Estimation of Gaussian Graphical Models. R package version 1.11. https://CRAN.R-project.org/package=glasso
}
\author{
Pol Castellano-Escuder
}
