% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaDensity.R
\name{PomaDensity}
\alias{PomaDensity}
\title{Density Plots}
\usage{
PomaDensity(
  data,
  x = "samples",
  outcome = NULL,
  feature_name = NULL,
  theme_params = list(legend_title = FALSE)
)
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{x}{Character. Options are "samples" (to visualize sample density plots) and "features" (to visualize feature density plots). Default is "samples".}

\item{outcome}{Character. Indicates the name of the \code{colData} column to be used as the outcome factor. Default is NULL (first factor variable in \code{colData}).}

\item{feature_name}{Character vector. Indicates the feature/s to display. Default is NULL (all features will be displayed).}

\item{theme_params}{List. Indicates \code{theme_poma} parameters.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{PomaDensity} generates a density plot for samples and features. This function can be used for data exploration (e.g., comparison between pre and post normalized datasets).
}
\examples{
data("st000284")

# Sample density plots
st000284 \%>\%
PomaNorm() \%>\% 
PomaDensity(theme_params = list(axistext = "y"))

# All feature density plots
st000284 \%>\% 
PomaNorm() \%>\% 
PomaDensity(x = "features", theme_params = list(legend_position = "none"))
             
# Specific feature density plots
st000284 \%>\% 
PomaNorm() \%>\% 
PomaDensity(x = "features", 
            feature_name = c("ornithine", "orotate"))
}
\author{
Pol Castellano-Escuder
}
