% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaDensity.R
\name{PomaDensity}
\alias{PomaDensity}
\title{Distribution Plot}
\usage{
PomaDensity(data, group = c("samples", "features"), feature_name = NULL)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{group}{Groupping factor for the plot. Options are c("samples", "features"). Option "samples" (default) will create a density plot for each group and option "features" will create a density plot of each variable.}

\item{feature_name}{A vector with the name/s of feature/s to plot. If it's NULL (default) a density plot of all variables will be created.}
}
\value{
A ggplot2 object.
}
\description{
PomaDensity() generates a density plot of not normalized and normalized MS data. This plot can help in the comparison between pre and post normalized data and in the "validation" of the normalization process.
}
\author{
Pol Castellano-Escuder
}
