% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaEDA.R
\name{PomaEDA}
\alias{PomaEDA}
\title{Automatic Exploratory Data Analysis HTML Report}
\usage{
PomaEDA(
  data,
  imputation = "knn",
  normalization = "log_pareto",
  clean_outliers = TRUE,
  coeff_outliers = 1.5,
  username = NULL,
  institution = NULL
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{imputation}{Imputation method. See \code{?POMA::PomaImpute()}.}

\item{normalization}{Normalization method. See \code{?POMA::PomaNorm()}.}

\item{clean_outliers}{Logical. If it's set to TRUE, outliers will be removed from EDA.}

\item{coeff_outliers}{This value corresponds to the classical 1.5 in \eqn{Q3 + 1.5*IQR} formula to detect outliers. See \code{?POMA::PomaOutliers()}.}

\item{username}{Author name in the report.}

\item{institution}{Institution name in the report.}
}
\value{
An exploratory data analysis HTML report.
}
\description{
This function automatically generates a HTML report with different exploratory plots and tables from an \code{SummarizedExperiment} object.
}
\author{
Pol Castellano-Escuder
}
