% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaHeatmap.R
\name{PomaHeatmap}
\alias{PomaHeatmap}
\title{Classical Heatmap}
\usage{
PomaHeatmap(data, sample_names = TRUE, feature_names = FALSE)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{sample_names}{Logical indicating if sample names should be plotted or not. Default is TRUE.}

\item{feature_names}{Logical indicating if feature names should be plotted or not. Default is FALSE.}
}
\value{
A heatmap.
}
\description{
This function returns a basic heatmap plot made with base R.
}
\author{
Pol Castellano-Escuder
}
