% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaHeatmap.R
\name{PomaHeatmap}
\alias{PomaHeatmap}
\title{Classical Heatmap}
\usage{
PomaHeatmap(data, scale = TRUE, scale_by = "features")
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{scale}{Logical that indicates if heatmap values should be scaled. Default is TRUE.}

\item{scale_by}{Variable to scale heatmap values. Options are "samples" and "features". Option "samples" (default) will \code{scale()} each sample considering all features and option "features" will \code{scale()} each feature considering all samples.}
}
\value{
A heatmap.
}
\description{
This function returns a basic heatmap plot made with base R.
}
\author{
Pol Castellano-Escuder
}
