% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaImpute.R
\name{PomaImpute}
\alias{PomaImpute}
\title{Collection of Imputation Methods for Mass Spectrometry Data}
\usage{
PomaImpute(
  data,
  ZerosAsNA = FALSE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = "knn"
)
}
\arguments{
\item{data}{A SummarizedExperiment object. First \code{colData} column must be the subject group/type.}

\item{ZerosAsNA}{Logical that indicates if the zeros in the data are missing values. Default is FALSE.}

\item{RemoveNA}{Logical that indicates if those features with more than selected cutoff missing values in each group have to be removed. Default is TRUE.}

\item{cutoff}{Numeric that indicates the percentage of missing values allowed in each group. If one of the groups have less missing values than selected cutoff value, these feature will not be removed.}

\item{method}{Imputation method. Options are: "none", "half_min", "median", "mean", "min", "knn" and "rf". If "none", all missing values will be replaced by zero.}
}
\value{
A SummarizedExperiment object with cleaned data.
}
\description{
PomaImpute() offers different methods to impute missing values in MS data.
}
\examples{
data("st000336")

PomaImpute(st000336, method = "knn")
}
\references{
Armitage, E. G., Godzien, J., Alonso‐Herranz, V., López‐Gonzálvez, Á., & Barbas, C. (2015). Missing value imputation strategies for metabolomics data. Electrophoresis, 36(24), 3050-3060.
}
\author{
Pol Castellano-Escuder
}
