% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaLimma.R
\name{PomaLimma}
\alias{PomaLimma}
\title{Easy Implementation of \code{limma} Bioconductor Package}
\usage{
PomaLimma(
  data,
  contrast = NULL,
  covariates = FALSE,
  covs = NULL,
  adjust = "fdr",
  weights = FALSE,
  cutoff = NULL
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{contrast}{A character with the limma comparison. For example, "Group1-Group2" or "control-intervention".}

\item{covariates}{Logical. If it's set to \code{TRUE} all metadata variables stored in \code{colData} will be used as covariables. Default = FALSE.}

\item{covs}{Character vector indicating the name of \code{colData} columns that will be included as covariates. Default is NULL (all variables).}

\item{adjust}{Multiple comparisons correction method. Options are: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH" and "BY".}

\item{weights}{Logical indicating whether the limma model should estimate the relative quality weights for each group. See \code{?limma::arrayWeights()}.}

\item{cutoff}{Default is NULL. If this value is replaced for a numeric value, the resultant table will contains only those features with an adjusted p-value below selected value.}
}
\value{
A tibble with limma results.
}
\description{
PomaLimma() uses the classical \code{limma} package.
}
\examples{
data("st000284")

st000284 \%>\%
  PomaNorm() \%>\%
  PomaLimma(contrast = "Healthy-CRC", adjust = "fdr")
}
\references{
Matthew E. Ritchie, Belinda Phipson, Di Wu, Yifang Hu, Charity W. Law, Wei Shi, Gordon K. Smyth, limma powers differential expression analyses for RNA-sequencing and microarray studies, Nucleic Acids Research, Volume 43, Issue 7, 20 April 2015, Page e47, https://doi.org/10.1093/nar/gkv007
}
\author{
Pol Castellano-Escuder
}
