% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaLM.R
\name{PomaLM}
\alias{PomaLM}
\title{Linear Models}
\usage{
PomaLM(data, x = NULL, y = NULL, adjust = "fdr")
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{x}{Character vector. Indicates the names of independent variables. If it's null, all features will be used.}

\item{y}{Character. Indicates the name of \code{colData} columns to be used as dependent variable. If it's set to NULL, the first numeric variable in \code{colData} will be used as the dependent variable.}

\item{adjust}{Character. Multiple comparisons correction method to adjust p-values. Available options are: "fdr" (false discovery rate), "holm", "hochberg", "hommel", "bonferroni", "BH" (Benjamini-Hochberg), and "BY" (Benjamini-Yekutieli).}
}
\value{
A \code{tibble} with the results.
}
\description{
\code{PomaLM} performs a linear model on a \code{SummarizedExperiment} object.
}
\examples{
data("st000284")

# Perform linear model with all features
st000284 \%>\% 
PomaLM()

# Perform linear model with two features
st000284 \%>\% 
PomaLM(x = c("x1_methyladenosine", "x2_deoxyuridine"))
}
\author{
Pol Castellano-Escuder
}
