% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaMSnSetClass.R
\name{PomaMSnSetClass}
\alias{PomaMSnSetClass}
\title{Convert data frames to a simplified MSnSet Object}
\usage{
PomaMSnSetClass(target, features)
}
\arguments{
\item{target}{Metadata variables structured in columns. Sample ID must be the first column and group/type/treatment of the study must be the second column.}

\item{features}{Table of features. Each feature in one column.}
}
\value{
A simplified MSnSet object.
}
\description{
Convert data frames to a simplified MSnSet Object
}
\references{
Laurent Gatto and Kathryn S. Lilley. MSnbase - an R/Bioconductor package for isobaric tagged mass spectrometry data visualization, processing and quantitation. Bioinformatics 28, 288-289 (2012).
}
\author{
Pol Castellano-Escuder
}
