% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaMultivariate.R
\name{PomaMultivariate}
\alias{PomaMultivariate}
\title{Multivariate Statistical Methods for Mass Spectrometry Data}
\usage{
PomaMultivariate(
  data,
  method = "pca",
  components = 5,
  center = FALSE,
  scale = FALSE,
  labels = FALSE,
  load_length = 1,
  ellipse = TRUE,
  validation = "Mfold",
  folds = 5,
  nrepeat = 10,
  vip = 1.5,
  num_features = 10,
  legend_position = "bottom"
)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{method}{A multivariate method. Options are: "pca", "plsda" and "splsda".}

\item{components}{Numeric. Number of components to include in the model. Default is 5.}

\item{center}{Logical that indicates whether the variables should be shifted to be zero centered. Default is FALSE.}

\item{scale}{Logical that indicates whether the variables should be scaled to have unit variance before the analysis takes place. Default is FALSE.}

\item{labels}{Logical indicating if sample names should be plotted or not.}

\item{load_length}{Numeric between 1 and 2. Define the length of biplot loadings. Default is 1.}

\item{ellipse}{Logical that indicates whether a 95\%CI ellipse should be plotted in scores plot. Default is TRUE.}

\item{validation}{(Only for "plsda" and "splsda" methods) Validation method. Options are "Mfold" and "loo".}

\item{folds}{(Only for "plsda" and "splsda" methods) Numeric. Number of folds for Mfold validation method (default is 5). If the validation method is loo, this value will become to 1.}

\item{nrepeat}{(Only for "plsda" and "splsda" methods) Numeric. Number of iterations for the validation method selected.}

\item{vip}{(Only for "plsda" method) Numeric indicating VIP cutoff to select features that will be displayed in vip plot.}

\item{num_features}{(Only for "splsda" method) Numeric. Number of variables selected to discriminate groups.}

\item{legend_position}{Character indicating the legend position. Options are "none", "top", "bottom", "left", and "right".}
}
\value{
A list with all results for multivariate statistical analysis including plots and data frames.
}
\description{
PomaMultivariate() allows users to perform different multivariate statistical analysis on MS data.
}
\examples{
data("st000336")

# PCA
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOutliers() \%>\%
  PomaMultivariate(method = "pca")

# PLSDA
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOutliers() \%>\%
  PomaMultivariate(method = "plsda", vip = 1)

# sPLSDA
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOutliers() \%>\%
  PomaMultivariate(method = "splsda")
}
\author{
Pol Castellano-Escuder
}
