% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaMultivariate.R
\name{PomaMultivariate}
\alias{PomaMultivariate}
\title{Multivariate Statistical Methods for Mass Spectrometry Data}
\usage{
PomaMultivariate(data_multi, method = c("pca", "plsda", "splsda"),
  components = 5, center = FALSE, scale = FALSE,
  validation = c("Mfold", "loo"), folds = 5, nrepeat = 10,
  num_features = 10)
}
\arguments{
\item{data_multi}{A MSnSet object. First \code{pData} column must be the suject group/type.}

\item{method}{A multivariate method. Options are c("pca", "plsda", "splsda").}

\item{components}{Numeric. Number of components to include in the model. Default is 5.}

\item{center}{Logical that indicates whether the variables should be shifted to be zero centered. Default is FALSE.}

\item{scale}{Logical that indicates whether the variables should be scaled to have unit variance before the analysis takes place. Default is FALSE.}

\item{validation}{(Only for "plsda" and "splsda" methods) Validation method. Options are c("Mfold", "loo").}

\item{folds}{(Only for "plsda" and "splsda" methods) Numeric. Number of folds for Mfold validation method (default is 5). If the validation method is loo, this value will become to 1.}

\item{nrepeat}{(Only for "plsda" and "splsda" methods) Numeric. Number of iterations for the validation method selected.}

\item{num_features}{(Only for "splsda" method) Numeric. Number of variables selected to discriminate groups.}
}
\value{
A list with all results for multivariate statistical analysis including plots and data frames.
}
\description{
PomaMultivariate() allows users to perform different multivariate statistical analysis on MS data.
}
\author{
Pol Castellano-Escuder
}
