% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaNorm.R
\name{PomaNorm}
\alias{PomaNorm}
\title{Collection of Normalization Methods for Mass Spectrometry Data}
\usage{
PomaNorm(data, method = "log_pareto", round = 3)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{method}{Normalization method. Options are: "none", "auto_scaling", "level_scaling", "log_scaling", "log_transformation", "vast_scaling" and "log_pareto".}

\item{round}{Numeric. Number of decimal places (Default is 3).}
}
\value{
A MSnSet object with normalized data.
}
\description{
PomaNorm() offers different methods to normalize MS data. This function contains both centering and scaling functions to normalize the data.
}
\examples{
data("st000284")

PomaNorm(st000284, method = "log_pareto")
}
\references{
van den Berg, R. A., Hoefsloot, H. C., Westerhuis, J. A., Smilde, A. K., & van der Werf, M. J. (2006). Centering, scaling, and transformations: improving the biological information content of metabolomics data. BMC genomics, 7(1), 142.
}
\author{
Pol Castellano-Escuder
}
