% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaNormPlot.R
\name{PomaNormPlot}
\alias{PomaNormPlot}
\title{Normalization Plot}
\usage{
PomaNormPlot(data, group = c("subjects", "features"))
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the suject group/type.}

\item{group}{Groupping factor for the plot. Options are c("subjects", "features"). If the user select "subject", the boxplot will be created for each subject. If the selection is "features", the boxplot will be created for each feature}
}
\value{
A ggplot2 object.
}
\description{
PomaNormPlot() generates a boxplot of not normalized and normalized MS data. This plot can help in the comparison between pre and post normalized data and in the "validation" of the normalization process.
}
\author{
Pol Castellano-Escuder
}
