% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaNormPlot.R
\name{PomaNormPlot}
\alias{PomaNormPlot}
\title{Normalization Plot}
\usage{
PomaNormPlot(data, group = c("samples", "features"), jitter = TRUE)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the suject group/type.}

\item{group}{Groupping factor for the plot. Options are c("samples", "features"). Option "samples" (default) will create a boxplot for each sample and option "features" will create a boxplot of each variable.}

\item{jitter}{Logical. If it's TRUE (default), the boxplot will show all points.}
}
\value{
A ggplot2 object.
}
\description{
PomaNormPlot() generates a boxplot of not normalized and normalized MS data. This plot can help in the comparison between pre and post normalized data and in the "validation" of the normalization process.
}
\author{
Pol Castellano-Escuder
}
