% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaOddsRatio.R
\name{PomaOddsRatio}
\alias{PomaOddsRatio}
\title{Logistic Regression Model Odds Ratios}
\usage{
PomaOddsRatio(
  data,
  feature_name = NULL,
  covariates = FALSE,
  covs = NULL,
  showCI = TRUE
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{feature_name}{A vector with the name/s of feature/s that will be used to fit the model. If it's NULL (default), all variables will be included in the model.}

\item{covariates}{Logical that indicates if covariates will be included in logistic regression model. Default is \code{FALSE}.}

\item{covs}{Character vector indicating the name of \code{colData} columns that will be included as covariates. Default is NULL (all variables).}

\item{showCI}{Logical that indicates if the 95\% confidence intervals will be plotted. Default is \code{TRUE}.}
}
\value{
A tibble with the Odds Ratios for all features with their 95\% confidence intervals and a ggplot2 object.
}
\description{
PomaOddsRatio() calculates the Odds Ratios for each feature from a logistic regression model using the binary outcome (group/type must be a binary factor) as a dependent variable.
}
\examples{
data("st000336")

st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOddsRatio(feature_name = c("glutamic_acid", "glutamine", 
                                 "glycine", "histidine"))
}
\author{
Pol Castellano-Escuder
}
