% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaOutliers.R
\name{PomaOutliers}
\alias{PomaOutliers}
\title{Remove and Analyze Outliers}
\usage{
PomaOutliers(
  data,
  do = "clean",
  method = "euclidean",
  type = "median",
  coef = 1.5,
  labels = FALSE
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{do}{Action to do. Options are "clean" (to remove detected outliers) and "analyze" (to analyze data outliers). Note that the output of this function will be different depending on this parameter.}

\item{method}{Distance measure method to perform MDS. Options are "euclidean", "maximum", "manhattan", "canberra" and "minkowski". See \code{?dist()}.}

\item{type}{Type of outliers analysis to perform. Options are "median" (default) and "centroid". See \code{vegan::betadisper}.}

\item{coef}{This value corresponds to the classical 1.5 in \eqn{Q3 + 1.5*IQR} formula to detect outliers. By changing this value, the permissiveness in outlier detection will change.}

\item{labels}{Logical indicating if sample IDs should to be plotted or not.}
}
\value{
A SummarizedExperiment object without outliers OR an exploratory outlier analysis including both plots and tables (depending on "do" parameter).
}
\description{
This function allows users to analyze outliers by different plots and remove them from an SummarizedExperiment object.
}
\examples{
data("st000336")

# clean outliers
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOutliers()

# analyze outliers
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOutliers(do = "analyze")
}
\author{
Pol Castellano-Escuder
}
