% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaPCA.R
\name{PomaPCA}
\alias{PomaPCA}
\title{Principal Components Analysis}
\usage{
PomaPCA(
  data,
  center = TRUE,
  scale = TRUE,
  ncomp = 4,
  labels = FALSE,
  ellipse = FALSE,
  load_length = 1,
  ...
)
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{center}{Logical. Indicates whether the variables should be shifted to be zero centered. Default is TRUE.}

\item{scale}{Logical. Indicates whether the variables should be scaled to have unit variance before the analysis takes place. Default is TRUE.}

\item{ncomp}{Numeric. Number of components to be included in the results. Default is 4.}

\item{labels}{Logical. Indicates if sample names should be displayed.}

\item{ellipse}{Logical. Indicates whether a 95 percent confidence interval ellipse should be displayed in score plot and biplot. Default is FALSE.}

\item{load_length}{Numeric. Indicates the length of biplot loading arrows. Value between 1 and 2. Default is 1.}
}
\value{
A \code{list} with results including plots and tables.
}
\description{
\code{PomaPCA} performs a principal components analysis on the given \code{SummarizedExperiment} object.
}
\examples{
data("st000336")

st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaPCA()
}
\author{
Pol Castellano-Escuder
}
