% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaPCR.R
\name{PomaPCR}
\alias{PomaPCR}
\title{Principal Components Regression}
\usage{
PomaPCR(
  data,
  n_components = 5,
  scale = TRUE,
  center = TRUE,
  outcome = NULL,
  intercept = TRUE
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{n_components}{The number of Principal Components used in the PCR model. Defaults is 5.}

\item{scale}{A logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place.}

\item{center}{A logical value indicating whether the variables should be shifted to be zero centered.}

\item{outcome}{Character string indicating the variable name in \code{colData} to be used as the outcome.}

\item{intercept}{A logical value indicating whether intercept should be included in the LM. Default is TRUE.}
}
\value{
A list with PCR results.
}
\description{
Fits a Linear Model using the indicated variable as the dependent variable (outcome) and the indicated number of Principal Components as independent variables.
}
\examples{
data("st000284")

st000284 \%>\%
  PomaPCR(outcome = "age_at_consent")

data("st000336")
  
st000336 \%>\% 
PomaImpute() \%>\% 
PomaPCR(n_components = 2, 
        outcome = "steroids", 
        intercept = FALSE)
}
\author{
Pol Castellano-Escuder
}
