% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaPCR.R
\name{PomaPCR}
\alias{PomaPCR}
\title{Principal Components Regression}
\usage{
PomaPCR(data, center = TRUE, scale = TRUE, ncomp = 2, y = NULL, adjust = "fdr")
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{center}{Logical. Indicates whether the variables should be shifted to be zero centered. Default is TRUE.}

\item{scale}{Logical. Indicates whether the variables should be scaled to have unit variance before the analysis takes place. Default is TRUE.}

\item{ncomp}{Numeric. Indicates the number of principal components used as predictors in the model. Default is 2.}

\item{y}{Character. Indicates the name of \code{colData} columns to be used as dependent variable. If it's set to NULL, the first numeric variable in \code{colData} will be used as the dependent variable.}

\item{adjust}{Character. Multiple comparisons correction method to adjust p-values. Available options are: "fdr" (false discovery rate), "holm", "hochberg", "hommel", "bonferroni", "BH" (Benjamini-Hochberg), and "BY" (Benjamini-Yekutieli).}
}
\value{
A \code{tibble} with the results.
}
\description{
\code{PomaPCR} performs Principal Components Regression.
}
\examples{
data("st000284")

PCR with 2 components
st000284 \%>\%
  PomaPCR(y = "age_at_consent")
  
PCR with 10 components
st000284 \%>\%
  PomaPCR(ncomp = 10)
}
\author{
Pol Castellano-Escuder
}
