% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaRandForest.R
\name{PomaRandForest}
\alias{PomaRandForest}
\title{Classification Random Forest}
\usage{
PomaRandForest(
  data,
  ntest = NULL,
  ntree = 500,
  mtry = floor(sqrt(ncol(t(SummarizedExperiment::assay(data))))),
  nodesize = 1,
  nvar = 20
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{ntest}{Numeric indicating the percentage of observations that will be used as test set. Default is NULL (no test set).}

\item{ntree}{Number of trees to grow.}

\item{mtry}{Number of variables randomly sampled as candidates at each split. This value is set sqrt(p) (where p is number of variables in data) by default.}

\item{nodesize}{Minimum size of terminal nodes. By default is equal to 1.}

\item{nvar}{Number of variables to show in the Gini plot.}
}
\value{
A list with all results for Random Forest including plots and tables.
}
\description{
PomaRandForest() allows users to perform a classification Random Forest using the classical \code{randomForest} R package.
}
\examples{
data("st000336")

st000336 \%>\% 
  PomaImpute() \%>\%
  PomaRandForest()
}
\references{
A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. R News 2(3), 18--22.
}
\author{
Pol Castellano-Escuder
}
