% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaUMAP.R
\name{PomaUMAP}
\alias{PomaUMAP}
\title{Dimensionality Reduction with UMAP}
\usage{
PomaUMAP(
  data,
  n_neighbors = NULL,
  n_components = 2,
  metric = "euclidean",
  pca = NULL,
  min_dist = 0.01,
  spread = 1,
  hdbscan_minpts = NULL,
  show_clusters = FALSE,
  show_group = FALSE,
  legend_position = "bottom"
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{n_neighbors}{The size of local neighborhood (sample points) used for manifold approximation.}

\item{n_components}{The dimension of the space to embed into. Defaults is 2.}

\item{metric}{Distance measure method to find nearest neighbors. Options are "euclidean", "cosine", "manhattan", "hamming" and "correlation". See \code{?uwot::umap()}.}

\item{pca}{If not NULL (default), reduce data to this number of columns using PCA before UMAP.}

\item{min_dist}{The effective minimum distance between embedded points.}

\item{spread}{The effective scale of embedded points.}

\item{hdbscan_minpts}{Integer; Minimum size of clusters. See \code{?dbscan::hdbscan()}.}

\item{show_clusters}{Logical indicating if clusters computed with HDBSCAN method should be plotted or not.}

\item{show_group}{Logical indicating if the original sample group from target should be plotted or not.}

\item{legend_position}{Character indicating the legend position. Options are "none", "top", "bottom", "left", and "right".}
}
\value{
A list with results including plots and tables.
}
\description{
Dimension reduction of the data using the Uniform Manifold Approximation and Projection (UMAP) method. See \code{?uwot::umap()} for more.
}
\examples{
data("st000284")

st000284 \%>\%
  PomaNorm() \%>\%
  PomaUMAP()
}
\references{
McInnes, L., Healy, J., & Melville, J. (2018). Umap: Uniform manifold approximation and projection for dimension reduction. arXiv preprint arXiv:1802.03426.

Campello, R. J., Moulavi, D., & Sander, J. (2013, April). Density-based clustering based on hierarchical density estimates. In Pacific-Asia conference on knowledge discovery and data mining (pp. 160-172). Springer, Berlin, Heidelberg.
}
\author{
Pol Castellano-Escuder
}
