% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaUnivariate.R
\name{PomaUnivariate}
\alias{PomaUnivariate}
\title{Univariate Statistical Methods for Mass Spectrometry Data}
\usage{
PomaUnivariate(
  data_uni,
  covariates = FALSE,
  method = c("ttest", "anova", "mann", "kruskal"),
  paired = FALSE,
  var_equal = FALSE,
  adjust = c("fdr", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY")
)
}
\arguments{
\item{data_uni}{A MSnSet object. First \code{pData} column must be the suject group/type.}

\item{covariates}{Logical. If it's set to \code{TRUE} all metadata variables stored in \code{pData} will be used as covariables. Default = FALSE.}

\item{method}{Univariate statistical method. Options are c("ttest", "anova", "mann", "kruskal").}

\item{paired}{Logical indicates if the data is paired or not.}

\item{var_equal}{Logical indicates if the data variance is equal or not.}

\item{adjust}{Multiple comparisons correction method.}
}
\value{
A data frame with results.
}
\description{
PomaUnivariate() allows users to perform different univariate statistical analysis on MS data.
}
\author{
Pol Castellano-Escuder
}
