% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaUnivariate.R
\name{PomaUnivariate}
\alias{PomaUnivariate}
\title{Univariate Statistical Methods for Mass Spectrometry Data}
\usage{
PomaUnivariate(
  data,
  covariates = FALSE,
  covs = NULL,
  method = "ttest",
  paired = FALSE,
  var_equal = FALSE,
  adjust = "fdr"
)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{covariates}{Logical. If it's set to \code{TRUE} all metadata variables stored in \code{colData} will be used as covariables. Default = FALSE.}

\item{covs}{Character vector indicating the name of \code{colData} columns that will be included as covariates. Default is NULL (all variables).}

\item{method}{Univariate statistical method. Options are: "ttest", "anova", "mann" and "kruskal".}

\item{paired}{Logical that indicates if the data is paired or not.}

\item{var_equal}{Logical that indicates if the data variance is equal or not.}

\item{adjust}{Multiple comparisons correction method. Options are: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH" and "BY".}
}
\value{
A tibble with results.
}
\description{
PomaUnivariate() allows users to perform different univariate statistical analysis on MS data.
}
\examples{
data("st000336")
data("st000284")

# ttest
st000336 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOutliers() \%>\%
  PomaUnivariate(method = "ttest")

# ANOVA
st000284 \%>\% 
  PomaImpute() \%>\%
  PomaNorm() \%>\%
  PomaOutliers() \%>\%
  PomaUnivariate(method = "anova")
}
\author{
Pol Castellano-Escuder
}
