% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaUnivariate.R
\name{PomaUnivariate}
\alias{PomaUnivariate}
\title{Title}
\usage{
PomaUnivariate(data_uni, covariates = NULL, method = c("ttest",
  "anova", "mann", "kruskal"), paired = FALSE, var_equal = FALSE,
  adjust = c("fdr", "holm", "hochberg", "hommel", "bonferroni", "BH",
  "BY"))
}
\arguments{
\item{data_uni}{A data frame}

\item{covariates}{A data frame}

\item{paired}{Logical that indicate if your data is paired or not}

\item{adjust}{}
}
\value{
A data.frame with
}
\description{
Title
}
