% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaUnivariate.R
\name{PomaUnivariate}
\alias{PomaUnivariate}
\title{Univariate Statistical Test}
\usage{
PomaUnivariate(
  data,
  method = "ttest",
  covs = NULL,
  error = NULL,
  paired_samples = FALSE,
  var_equal = FALSE,
  adjust = "fdr",
  run_post_hoc = TRUE
)
}
\arguments{
\item{data}{A \code{SummarizedExperiment} object.}

\item{method}{Character. The univariate statistical test to be performed. Available options include "ttest" (T-test), "anova" (analysis of variance), "mann" (Wilcoxon rank-sum test), and "kruskal" (Kruskal-Wallis test).}

\item{covs}{Character vector. Indicates the names of \code{colData} columns to be included as covariates. Default is NULL (no covariates). If not NULL, an ANCOVA model will be fitted using the specified covariates. Note: The order of the covariates is important and should be listed in increasing order of importance in the experimental design.}

\item{error}{Character vector. Indicates the name of a \code{colData} column to be included as an error term (e.g. replicates). Default is NULL (no error term).}

\item{paired_samples}{Logical. Indicates if the data is paired or not. Default is FALSE.}

\item{var_equal}{Logical. Indicates if the data variances are assumed to be equal or not. Default is FALSE.}

\item{adjust}{Character. Multiple comparisons correction method to adjust p-values. Available options are: "fdr" (false discovery rate), "holm", "hochberg", "hommel", "bonferroni", "BH" (Benjamini-Hochberg), and "BY" (Benjamini-Yekutieli).}

\item{run_post_hoc}{Logical. Indicates if computing post-hoc tests or not. Setting this parameter to FALSE can save time for large datasets.}
}
\value{
A \code{list} with the results.
}
\description{
\code{PomaUnivariate} performs parametric and non-parametric univariate statistical tests on a \code{SummarizedExperiment} object to compare groups or conditions. Available methods include T-test, ANOVA, ANCOVA, Mann Whitney U Test (Wilcoxon Rank Sum Test), and Kruskal-Wallis.
}
\examples{
data("st000336")

# Perform T-test
st000336 \%>\% 
PomaImpute() \%>\% 
PomaUnivariate(method = "ttest")

# Perform Mann-Whitney U test
st000336 \%>\% 
PomaImpute() \%>\% 
PomaUnivariate(method = "mann", paired_samples = FALSE, adjust = "fdr")

data("st000284")
# Perform Two-Way ANOVA
st000284 \%>\% 
PomaUnivariate(method = "anova", covs = c("gender"))

# Perform Three-Way ANOVA
st000284 \%>\% 
PomaUnivariate(method = "anova", covs = c("gender", "smoking_condition"))

# Perform ANCOVA with one numeric covariate and one factor covariate
# st000284 \%>\% 
# PomaUnivariate(method = "anova", covs = c("age_at_consent", "smoking_condition"))

# Perform Kruskal-Wallis test
st000284 \%>\% 
PomaUnivariate(method = "kruskal", adjust = "holm")
}
\author{
Pol Castellano-Escuder
}
