% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaUnivariate.R
\name{PomaUnivariate}
\alias{PomaUnivariate}
\title{Univariate Statistical Methods for Mass Spectrometry Data}
\usage{
PomaUnivariate(
  data,
  covariates = FALSE,
  method = "ttest",
  paired = FALSE,
  var_equal = FALSE,
  adjust = "fdr"
)
}
\arguments{
\item{data}{A MSnSet object. First \code{pData} column must be the subject group/type.}

\item{covariates}{Logical. If it's set to \code{TRUE} all metadata variables stored in \code{pData} will be used as covariables. Default = FALSE.}

\item{method}{Univariate statistical method. Options are: "ttest", "anova", "mann" and "kruskal".}

\item{paired}{Logical indicates if the data is paired or not.}

\item{var_equal}{Logical indicates if the data variance is equal or not.}

\item{adjust}{Multiple comparisons correction method. Options are: "fdr", "holm", "hochberg", "hommel", "bonferroni", "BH" and "BY".}
}
\value{
A data frame with results.
}
\description{
PomaUnivariate() allows users to perform different univariate statistical analysis on MS data.
}
\author{
Pol Castellano-Escuder
}
