% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomaVolcano.R
\name{PomaVolcano}
\alias{PomaVolcano}
\title{Volcano Plot}
\usage{
PomaVolcano(
  data,
  pval = c("raw", "adjusted"),
  Pval_cutoff = 0.05,
  FC_cutoff = 1.5,
  xlim = 2
)
}
\arguments{
\item{data}{A data frame object from PomaUnivariate(method = c("ttest", "mann")). Only for two group data!}

\item{pval}{Select a pvalue type to generate the volcano plot. Options c("raw", "adjusted").}

\item{Pval_cutoff}{Numeric. Define the pvalue cutoff (horizontal line).}

\item{FC_cutoff}{Numeric. Define the fold change cutoff (vertical lines).}

\item{xlim}{Numeric. Define the limits for x axis.}
}
\value{
A ggplot2 object.
}
\description{
PomaVolcano() generates a volcano plot from the PomaUnivariate() result. The data can't has negative values, so select method = "none" in PomaNorm() to avoid them.
}
\author{
Pol Castellano-Escuder
}
