% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_poma_c}
\alias{scale_fill_poma_c}
\title{Fill scale constructor for continuous \code{viridis} "plasma" palette}
\usage{
scale_fill_poma_c()
}
\description{
Fill scale constructor for continuous \code{viridis} "plasma" palette
}
